feat:exception handling

+++ src/main/java/com/Ecommerce/hicommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package com.Ecommerce.hicommerce.services;
 import com.Ecommerce.hicommerce.dto.ProductDTO;
 import com.Ecommerce.hicommerce.entities.Product;
 import com.Ecommerce.hicommerce.repositories.ProductRepository;
+import com.Ecommerce.hicommerce.services.exceptions.DatabaseException;
 import com.Ecommerce.hicommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
@@ -45,15 +49,29 @@ public class ProductService {
     @Transactional
     public  ProductDTO update (Long id , ProductDTO productDTO){
 
-        Product product = productRepository.getReferenceById(id);
-        copyDtoToEntity(productDTO,product);
-        product = productRepository.save(product);
-        return new ProductDTO(product);
+        try {
+            Product product = productRepository.getReferenceById(id);
+            copyDtoToEntity(productDTO,product);
+            product = productRepository.save(product);
+            return new ProductDTO(product);
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Product not found");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public  void delete (Long id ){
-         productRepository.deleteById(id);
+        if(!productRepository.existsById(id)){
+            throw new ResourceNotFoundException("Product not found");
+        }
+        try {
+            productRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
+
     }
 
     private void copyDtoToEntity(ProductDTO productDTO, Product product) {
